Add 2 LUTs for HI use
The first is for the HI centrality trigger and is already populated with a first draft of the LUT contents. The second is for the future use of the Q2 trigger and is not populated yet.

This also involved changes to the conversion to the legacy dataformats that were worked out with the HI L1 menu definition.