ReFactor calospares paths
The CaloSpares code has been refactored such that the Stage 1 emulator now produces two collections, one for HFRingSums, and another for HFBitCounts. Previously they were the same collection marked with an enum to differentiate them. The change was prompted by conversations with the unpacker/packer group because of the possibility of having two emulator outputs in the same HFRingSum (Q2 and centrality trigger