Consumes migration for L1GtUtils
This modifies L1GtUtils to call the consumes function
for all the products it gets. Also the L1GtAnalyzer
module has consumes calls added. They also use
getByToken instead of getByLabel now.

I did not yet modify all the modules that use L1GtUtils
yet. This will be done next in a separate commit.

Modules will need to use a different L1GtUtils
constructor and the arguments determine the behavior.
This preserves the previous behavior where it is
possible to have L1GtUtils look for the products
by type without specifying the InputTags in the
configuration. This will make the migration of the
modules easier as configurations will not need
modification to add the InputTag parameters.
It is also possible to pass the InputTags into the
constructor as arguments which was also an
option in the earlier versions. And I added the
the new possibility to set the InputTags in the
configuration as well.

All modules that use L1GtUtils will also have to
remove InputTags where they were being passed in
at points other than the constructor. They have to
be set in the constructor now. This has to
be decided earlier and the module can no longer
decide the InputTag when processing the event or
run.

The new L1GtUtilsHelper class encapsulates most
of the logic of the input tag selection and
holds the EDGetTokens.