Input tag for L1GlobalTriggerReadoutRecord from provenance.
Remove the arbitrariness of choosing the input tag for
L1GlobalTriggerReadoutRecord product when the input tag is retrieved
from provenance and more products of this type exist, with different
input tags. 

The code will look for the corresponding product in ProductRegistry.
If the product is found, it checks the product label in a vector of
preferred input tags (hardwired now to "gtDigis" and "hltGtDigis"). 

The first input tag from the vector of preferred input tags, with the
same label as the input tag found from provenance, is kept as input tag,
if there are no multiple products with the same label.

If multiple products are found and no one has a label in the vector of
preferred input tags, or if multiple products are found with the label
in the vector of preferred input tags (with different instance or
process) the input tag is set to empty input tag, and L1GtUtil will
produce an error, as it is not possible to safely choose a product. In
this case, one must provide explicitly the correct input tag via
configuration or in the constructor.


The same treatment is applied to L1GlobalTriggerRecord and
L1GtTriggerMenuLite.