Modernized use of eras and fixed problem with fastSim configuration
The 'isChosen' interface is deprecated for Modifiers so needed to
be replaced with with modern recommended usage for eras.

The module l1L1GtObjectMap never worked properly when run in fastSim
since it needed data from a module which was run in a later Path.
The result was the data from l1L1GtObjectMap was never put into the
event for fastSim. Now that the framework is going to run Paths
in parallel, the improper data dependency is producing an unrunnable
schedule. Therefore remove the module from the Sequence gives
functionally equivalent behavior as before this change.