Replaced tab with space to avoid gcc 6.3 warning in L1Trigger/L1TCommon
Because of the different possible spacing used for tabs, gcc 6.3
was warning of a line after an `if` that might appear to visually
be part of the `if` but is, in fact, not. Switching from tabs to
using spaces, which is consistent with the rest of the document,
removes the warning.