Fix includes of PrimitiveConverter_Neighbor.h
We use CSCDetId in the metohdPrimConv, so we also need to include
the CSCDetID.h header to make this file compile on its own.

We also use L1TMuon::TriggerPrimitive in the same method, so
we need to include the relevant header for the same reason.

I also qualified that the TriggerPrimitive we use in this file
is actually in the L1TMuon namespace.