Replaced use of sprintf with efficient use of LogTrace
The code was using sprintf to fill a char array which was only
being passed to LogTrace. gcc 8 gave an error that the buffer being
passed to sprintf was insufficient to guarantee that there would be
no memory overwrites. Switching to just using LogTrace directly
and adding a conditional compilation avoids the possible overwrite
and should improve performance.